pyproject.toml
[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"


[tool.poetry]
name = "platonic-website"
version = "0.1.0"
description = "Documentation website for python-platonic project"
license = "MIT"

authors = []

readme = "README.md"

repository = "https://github.com/python-platonic/platonic-website"

keywords = []

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "3.8"
mkdocs = "^1.1.2"
mkdocs-material = "^6.1.0"
mkautodoc = "^0.1.0"
mkdocs-macros-plugin = "^0.5.0"
platonic-sqs = "^1.2.4"

[tool.poetry.dev-dependencies]
mister-jeeves = "^0.0.2"
